Fixed QR Decomposition

+++ utils.h
@@ -8,7 +8,7 @@ void vectorPrint(float* u , int rows);
 
 void vectorScale(float* u, int rows, float alpha , float* v); 
 
-int vectorDotProduct(float* u, float* v, int rowsU, int rowsV);
+float vectorDotProduct(float* u, float* v, int rowsU, int rowsV);
 
 void vectorSubtract(float* u, int rowsU, float* v, int rowsV, float* z);
